+++ develop/src/Base64.cpp
@@ -101,7 +101,7 @@ std::string encode(std::string const& data)
 
 std::string decode(std::string const& encoded_string)
 {
-    int in_len = encoded_string.size();
+    int in_len = (int)encoded_string.size();
     int i = 0;
     int j = 0;
     int in_ = 0;
